LayoutTests/fast/events/onunload-window-property-expected.txt

main frame "1" - has 1 onunload handler(s)
ALERT: unload
You should have seen an unload alert appear.
